# backend/docs/swagger.yaml
basePath: /
definitions:
...
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          details:
            items:
              $ref: '#/definitions/utils.ErrorDetail'
            type: array
          message:
            type: string
          request_id:
            type: string
          timestamp:
            type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://example.com/support
  description: API for Reelcut — clip creation from long-form video (auth, projects,
    videos, transcriptions, analysis, clips, templates, jobs).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: Reelcut API
  version: "1.0"
paths:
  /api/v1/analysis/videos/{videoId}:
    get:
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get analysis by video ID
      tags:
      - analysis
    post:
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start analysis for a video
      tags:
      - analysis
...
  /api/v1/transcriptions/videos/{videoId}:
    get:
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transcription by video ID
      tags:
      - transcriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      - description: language, enable_diarization
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start transcription for a video
      tags:
      - transcriptions
...
  /api/v1/videos:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Filter by project
        in: query
        name: project_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: created_at
        in: query
        name: sort_by
        type: string
      - description: asc|desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List videos
      tags:
      - videos
  /api/v1/videos/{id}:
    delete:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a video
      tags:
      - videos
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get video by ID
      tags:
      - videos
  /api/v1/videos/{id}/confirm:
    post:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Confirm video upload completed
      tags:
      - videos
  /api/v1/videos/{id}/metadata:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get video metadata
      tags:
      - videos
  /api/v1/videos/{id}/thumbnail:
    get:
      parameters:
      - description: Video ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "302":
          description: Redirect to thumbnail
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Redirect to video thumbnail URL
      tags:
      - videos
  /api/v1/videos/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: project_id, filename
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get presigned URL for video upload
      tags:
      - videos
  /api/v1/videos/upload/url:
    post:
      responses:
        "501":
          description: Not Implemented
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Upload video from URL (not implemented)
      tags:
      - videos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"